services/producer/requirements.txt
confluent-kafka==2.5.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
